fix(codex): keep steering image contracts acyclic
Keep image-order types as a dependency leaf and register the stored-image regression suite in its canonical CI shard.
