+++ src/auto-reply/reply/reply-run-registry.message-injection.ts
@@ -1,5 +1,5 @@
 import { normalizeOptionalString } from "@openclaw/normalization-core/string-coerce";
-import { hasPromptImageInput } from "../../media/prompt-image-order.js";
+import { hasPromptImageInput } from "../../media/prompt-image-input.js";
 import { createDeferredCore } from "../../shared/deferred.js";
 import {
   replyMessageInjectionTargetOperation,
