+++ src/plugin-sdk/session-transcript-runtime.ts
@@ -54,7 +54,7 @@ export type {
   TranscriptEntryAnchor,
   TranscriptTurnAdmission,
 } from "../config/sessions/session-accessor.js";
-export { hasPromptImageInput } from "../media/prompt-image-order.js";
+export { hasPromptImageInput } from "../media/prompt-image-input.js";
 
 export {
   formatSessionTranscriptMemoryHitKey,
